Skybox textures for World 1
Updated software; 27 islands all populated with their own set of
foliage, 2 layers of 3D volumetric clouds, the shader used for the
islands is made up of three materials, the shaders used for the clouds
are made up of two materials, everything in the scene was ray-traced
with the global illumination settings and in order to not have seams I
have to do a pre-pass render for the lighting meaning, 11hr+ total
render time for all 6 shots. Hope you enjoy!